requirements.txt
absl-py==2.3.1
alembic==1.14.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.8
edward2==0.0.2
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.4.0
gin-config==0.5.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
gym==0.23.0
gym-notices==0.1.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Mako==1.3.10
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
optuna==4.4.0
packaging==25.0
pandas==2.0.3
pillow==10.4.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pygame==2.1.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
recsim-ng==0.1.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.10.1
six==1.17.0
SQLAlchemy==2.0.24
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.20.1
termcolor==2.4.0
tf-agents==0.17.0
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.14.1
zipp==3.20.2
